fix dependency cycle
