+++ packages/app/features/browser/BrowserCredentialHandoffScreen.tsx
@@ -1,17 +1,8 @@
 import { Button, Icon, Pressable, Text } from '@tloncorp/ui';
-import {
-  createContext,
-  type PropsWithChildren,
-  useCallback,
-  useContext,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
+import { useCallback, useEffect, useState } from 'react';
 import { View, XStack, YStack } from 'tamagui';
 
-import type { BrowserCredentialHandoffParams } from '../../navigation/BasePathNavigator';
+import type { BrowserCredentialHandoffParams } from '../../navigation/types';
 import {
   Field,
   ScreenHeader,
@@ -23,69 +14,7 @@ import {
   beginBrowserCredentialHandoff,
   submitBrowserCredentials,
 } from './browserCredentialHandoff';
-
-type CompletionHandler = () => Promise<void>;
-
-type BrowserCredentialHandoffCompletionContextValue = {
-  register: (handler: CompletionHandler) => string;
-  complete: (id: string) => Promise<void>;
-  discard: (id: string) => void;
-};
-
-const BrowserCredentialHandoffCompletionContext =
-  createContext<BrowserCredentialHandoffCompletionContextValue | null>(null);
-
-export function BrowserCredentialHandoffCompletionProvider({
-  children,
-}: PropsWithChildren) {
-  const handlers = useRef(new Map<string, CompletionHandler>());
-  const sequence = useRef(0);
-
-  const register = useCallback((handler: CompletionHandler) => {
-    const id = `browser-handoff-${Date.now()}-${++sequence.current}`;
-    handlers.current.set(id, handler);
-    return id;
-  }, []);
-
-  const complete = useCallback(async (id: string) => {
-    const handler = handlers.current.get(id);
-    if (!handler) {
-      throw new Error('The originating conversation is no longer available.');
-    }
-    handlers.current.delete(id);
-    try {
-      await handler();
-    } catch (error) {
-      handlers.current.set(id, handler);
-      throw error;
-    }
-  }, []);
-
-  const discard = useCallback((id: string) => {
-    handlers.current.delete(id);
-  }, []);
-
-  const value = useMemo(
-    () => ({ register, complete, discard }),
-    [complete, discard, register]
-  );
-
-  return (
-    <BrowserCredentialHandoffCompletionContext.Provider value={value}>
-      {children}
-    </BrowserCredentialHandoffCompletionContext.Provider>
-  );
-}
-
-export function useBrowserCredentialHandoffCompletion() {
-  const value = useContext(BrowserCredentialHandoffCompletionContext);
-  if (!value) {
-    throw new Error(
-      'Browser credential handoff completion provider is unavailable.'
-    );
-  }
-  return value;
-}
+import { useBrowserCredentialHandoffCompletion } from './BrowserCredentialHandoffCompletion';
 
 type Props = {
   navigation: { goBack(): void };
